move solution filtering to domain
